backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-decouple